bugfix: radar chart line does not respect chartColors
Option chartColors sets the color of markers on a radar chart but is
ignored for lines which take Powerpoint's default colors.

In case there is no dataBorder specify, fallback to chartColors.

This is very similar to bubble chart fallback (which is reindented to be
be more explicit in the XML elements it creates) except that cap can be
parameterized in radar.